Add color legend to VistA Menus viz.
Add the code necessary to capture the "TYPE" of each entry in the OPTION
file.  This will allow the Viz to denote what type of option is available
within each menu.

update the VistAMenu json files with the new information.

Add a menutype json to keep the color, internal name, and display name.