Add imageCompare test to VistA Menus
Using the PILLOW library in python:
  http://pillow.readthedocs.org/en/3.0.x/index.html

generate an image comparison test for the menu highlighting capability of
the VistA Menus page.  Two images should be expected to fail and those
difference images should be sent to the Dashboard.

The image will fail if
there are >3000 pixels different between the new image and the basline