Add Protocol information to Package information
Capture the rest of the Protocol information for all of the packages and
add a link to the protocols in the "Interfaces" accordion for each package.

Requires that the backend informatio be generated  an update to the
VistADataParse branch. The fork can be found here and will be merged to
the VistADataParse branch soon:

https://github.com/josephsnyder/VistA/tree/protocol_generation_update