Add visualization for INSTALL file
Add visualization which display the time order of the installed updates to
a VistA package.  The size of each bar in the timeline represents the
amount of fields and files that were installed.

This install information was gathered from the February 2016 FOIA release.

The information for this update is generated by update to the
VistADataParse branch which can be found at:

https://github.com/josephsnyder/VistA/tree/add_install_json_generation

Future Work: Allow for searching? Show multiple packages at once?