Add Patch Dependency Visualization
Add another TreeView visualization which takes the package autocomplete
and the install_autocomplete to show an install's dependencies.

It should trace a patch install dependency for as far as possible, with
the current information that is available.  In the end a new method may be
necessary to show all patch dependencies.