Update VistA Menu highlight and selection
Add a hover trigger to show the parent menu's name when the mouse is over
an entry in the option_autocomplete.  This will allow the user to
differentiate between multiple entries of the same option.

Change the option_autocomplete to only show one entry per top-level menu
and change the highlighting to show paths to all entries of that option
within the menu