Add testing for updates to ViViaN menu display
Add tests for the new functionality to the ViViaN menu/option display
page. Add tests for the panning/zooming feature which were adapted from
the patch dependency tests.

Add a second image compare which features multiple paths of menus highlighted.

Add a test which ensures that the top level menu text is shown when an
option autocomplete entry is hovered over.