Update PackageCategories.json
Noted that I created the following fields:
* “name”
* “description”
* “vdl”
* “status”
* “gui”
* “gui-vdl”
* “primary-function”


The gui-vdl is a separate vdl link specifically for the gui documentation.

The primary-function is either yes or no.  “Yes” means the function
described provides the primary function for the package.  Example is the
“lab service” which has the primary function to “Provide General
Laboratory Services”, but it also provides secondary function for “Provide
Pathology Services”

The only other case is the ADT vs. Registration functions which use the
Registration package.