Add Dependent Patch information
In addition to the original information of patch dependency, add a
toggle to read in patch dependent information.  This will transform the
table to show as "children", the patches which depend upon that patch to
exist.

Ensure the switching mechanism resets both sets of information within
the autocomplete.

Account for patches that may not be depended upon by anything.

The newly implemented arrow head on each line will always point to the dependency.